Udp parsing working
Port when the ports are udp was not working.
It seems the method tabular.GetColumnNoHeader doesn't not have the same
behaviour with tcp or udp ports, so we have implemented GetAllNoHeader
that gives all the data without the header.
It works because to find the ports it's implemented with a regular
expression that it works finding in all the text